SDR-331 refactor 패키지 이름 변경 aws -> image
