+++ be/src/main/java/com/jjikmuk/sikdorak/image/api/AwsCommandApi.java
@@ -1,12 +1,12 @@
-package com.jjikmuk.sikdorak.aws.api;
+package com.jjikmuk.sikdorak.image.api;
 
-import com.jjikmuk.sikdorak.aws.command.app.GeneratePreSignedURLService;
-import com.jjikmuk.sikdorak.aws.command.app.ImageMetaDataService;
-import com.jjikmuk.sikdorak.aws.command.app.request.PreSignedUrlCreateRequest;
-import com.jjikmuk.sikdorak.aws.command.app.response.PreSignedUrlCreateResponse;
 import com.jjikmuk.sikdorak.common.ResponseCodeAndMessages;
 import com.jjikmuk.sikdorak.common.aop.UserOnly;
 import com.jjikmuk.sikdorak.common.response.CommonResponseEntity;
+import com.jjikmuk.sikdorak.image.command.app.GeneratePreSignedURLService;
+import com.jjikmuk.sikdorak.image.command.app.ImageMetaDataService;
+import com.jjikmuk.sikdorak.image.command.app.request.PreSignedUrlCreateRequest;
+import com.jjikmuk.sikdorak.image.command.app.response.PreSignedUrlCreateResponse;
 import com.jjikmuk.sikdorak.user.auth.api.AuthenticatedUser;
 import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
 import lombok.RequiredArgsConstructor;
