+++ be/src/main/java/com/jjikmuk/sikdorak/image/command/app/ImageMetaDataService.java
@@ -1,6 +1,6 @@
-package com.jjikmuk.sikdorak.aws.command.app;
+package com.jjikmuk.sikdorak.image.command.app;
 
-import com.jjikmuk.sikdorak.aws.command.domain.ImageMetaData;
+import com.jjikmuk.sikdorak.image.command.domain.ImageMetaData;
 import com.jjikmuk.sikdorak.user.auth.api.LoginUser;
 
 public interface ImageMetaDataService {
