+++ be/src/main/java/com/jjikmuk/sikdorak/image/exception/InvalidImagesExtensionException.java
@@ -1,4 +1,4 @@
-package com.jjikmuk.sikdorak.aws.exception;
+package com.jjikmuk.sikdorak.image.exception;
 
 import com.jjikmuk.sikdorak.common.exception.SikdorakRuntimeException;
 import org.springframework.http.HttpStatus;
